+++ apps/api/src/chat/turn-runner.ts
@@ -205,7 +205,7 @@ export const runChatSessionTurn = async (input: RunChatSessionTurnInput): Promis
 		{ buildSubmitDiagnosisTool },
 		{ McpToolExecutor },
 	] = await Promise.all([
-		import("../app"),
+		import("../runtime/service-graph"),
 		import("../platform/DatabasePgLive"),
 		import("../platform/Llm"),
 		import("./loop"),
